Hunyuan1.5 (#484)
Thanks to HunyuanVideo Team and ModelTC Team.

---------

+++ lightx2v/models/input_encoders/hf/q_linear.py
@@ -162,7 +162,7 @@ class SglQuantLinearFp8(nn.Module):
             input_tensor_quant,
             self.weight.t(),
             input_tensor_scale,
-            self.weight_scale,
+            self.weight_scale.float(),
             dtype,
             bias=self.bias,
         )
@@ -249,7 +249,7 @@ class Q8FQuantLinearInt8(nn.Module):
         output_tensor = q8_linear(
             input_tensor_quant,
             self.weight,
-            self.bias if self.bias is not None else None,
+            self.bias.float() if self.bias is not None else None,
             input_tensor_scale,
             self.weight_scale.float(),
             fuse_gelu=False,
@@ -295,9 +295,9 @@ class Q8FQuantLinearFp8(nn.Module):
         output_tensor = fp8_linear(
             input_tensor_quant,
             self.weight,
-            self.bias if self.bias is not None else None,
+            self.bias.float() if self.bias is not None else None,
             input_tensor_scale,
-            self.weight_scale,
+            self.weight_scale.float(),
             out_dtype=torch.bfloat16,
         )
         return output_tensor
